+++ src/commands/vscode-command-data.ts
@@ -1,13 +1,15 @@
-import { Uri, Webview } from "vscode";
+import { SecretStorage, Uri, Webview } from "vscode";
 
 export class VsCodeCommandData<T> {
     webview: Webview;
     extensionUri: Uri;
     payload: T;
+    secrets: SecretStorage;
 
-    constructor(webview: Webview, extensionUri: Uri, payload: T) {
+    constructor(webview: Webview, extensionUri: Uri, payload: T, secrets: SecretStorage) {
         this.webview = webview;
         this.extensionUri = extensionUri;
         this.payload = payload;
+        this.secrets = secrets;
     }
 }